fix bug:SRAI used error funct7,lead to virtio writed error value of avail ring and used ring to emulator

+++ src/virtio_blk.h
@@ -53,7 +53,7 @@ LIST_HEAD(disk_op_list, disk_operation);
 
 void virtio_blk_init(const char *disk_image_path);
 uint32_t virtio_mmio_read(void *opaque,uint64_t offset,uint8_t size);
-void virtio_mmio_write(void *opaque,uint64_t offset, uint32_t value,uint8_t size) ;
+void virtio_mmio_write(void *opaque,uint64_t offset, uint64_t value,uint8_t size) ;
 void virtio_blk_raise_interrupt(void);  
 void virtio_disk_update(uint64_t *current_cycle);
 #endif
